release 3.1.70 source code

+++ services/bss/src/main/java/com/huaweicloud/sdk/bss/v2/model/CreatePartnerCouponsReq.java
@@ -3,6 +3,7 @@ package com.huaweicloud.sdk.bss.v2.model;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
@@ -26,7 +27,7 @@ public class CreatePartnerCouponsReq {
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "face_value")
 
-    private Double faceValue;
+    private BigDecimal faceValue;
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "valid_time")
@@ -108,7 +109,7 @@ public class CreatePartnerCouponsReq {
         this.customerIds = customerIds;
     }
 
-    public CreatePartnerCouponsReq withFaceValue(Double faceValue) {
+    public CreatePartnerCouponsReq withFaceValue(BigDecimal faceValue) {
         this.faceValue = faceValue;
         return this;
     }
@@ -119,11 +120,11 @@ public class CreatePartnerCouponsReq {
      * maximum: 99999999999
      * @return faceValue
      */
-    public Double getFaceValue() {
+    public BigDecimal getFaceValue() {
         return faceValue;
     }
 
-    public void setFaceValue(Double faceValue) {
+    public void setFaceValue(BigDecimal faceValue) {
         this.faceValue = faceValue;
     }
